rename exception.h to openraveexception.h
