+++ include/openrave/openraveexception.h
